lib/Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pymongo==4.10.1
google-generativeai==0.7.2
youtube-transcript-api==0.6.2
gTTS==2.5.3
